requirements.txt
flask
pymysql
pycryptodome
python-dotenv
gunicorn